[INJIMOB-843]: refactor the screenHeight and isSmallscreen into a common function
