+++ .talismanrc
@@ -213,8 +213,10 @@ fileignoreconfig:
     - filename: screens/Home/MyVcs/IdInputModal.tsx
     checksum: c18e8bbcf1c8de3327c0920fddcf1b6df54b74d894e3f04677f498712c680fe9  
   - filename: screens/Home/MyVcs/OtpVerificationModal.tsx
-    checksum: e1209eeda11c5c669b04b7a68674593dc4280945274b1598d5d0f9345ba33930  
+    checksum: c42e2387dfc2cc637033ae849d8b2c2fc30cabaeb77e074661cf466a8476e947  
+  - filename: shared/commonUtil.ts
+    checksum: f98c2c6a9add155155d11b231b2c5db18eb0b4570990990c8300e621609e2c78 
   - filename: screens/Home/MyVcs/GetIdInputModal.tsx
-    checksum: 52bdb70fe99eb2fcde949980c100953f5b26a5a98b5d9825db425331b9a5311c  
+    checksum: 1a15e4327b358ffee8bddfe412a38873ad6d4414d2b1d6bffd3c1782f5ec852f  
   version: ""
